section 9.108 - jpa delete by id not found

+++ src/main/java/guru/springframework/spring6restmvc/services/BeerServiceJPA.java
@@ -57,9 +57,12 @@ public class BeerServiceJPA implements BeerService {
     }
 
     @Override
-    public void deleteBeerById(UUID beerId) {
-        System.out.println("JPA Impl - deleting by id.");
-        beerRepository.deleteById(beerId);
+    public Boolean deleteBeerById(UUID beerId) {
+        if (beerRepository.existsById(beerId)) {
+            beerRepository.deleteById(beerId);
+            return true;
+        }
+        return false;
     }
 
     @Override
